refactor : 전반적인 파일 구조 변경(각각의 주제(ex.lineManager,stationManager)에서 기능(view,manager)으로)
