+++ src/component/text.js
